fix(adk): preserve nil agentCancelOpts in stopSignal.check to prevent UntilIdleFor from canceling agent (#972)
